+++ sdk/core/azure-core/azure/core/pipeline/policies/_retry.py
@@ -181,9 +181,13 @@ class RetryPolicy(HTTPPolicy):
         :rtype: int
         """
         retry_after = response.http_response.headers.get("Retry-After")
-        if retry_after is None:
-            return None
-        return self.parse_retry_after(retry_after)
+        if retry_after:
+            return retry_after
+        retry_after = response.http_response.headers.get("retry-after-ms")
+        if retry_after:
+            retry_after = retry_after / 1000
+            return retry_after
+        return None
 
     def _sleep_for_retry(self, response, transport):
         """Sleep based on the Retry-After response header value.